+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Stop transaction attempts when prerequisite synchronization fails or is cancelled [#953].
+- Prevent pending login migration or password setup from restoring reset wallet credentials [#953].
+
 ## [1.8.0] - 2026-03-25
 
 ### Changed
@@ -671,6 +674,7 @@ dusk-network/rusk#3984: https://github.com/dusk-network/rusk/issues/3984
 [#939]: https://github.com/dusk-network/web-wallet/issues/939
 [#946]: https://github.com/dusk-network/web-wallet/issues/946
 [#948]: https://github.com/dusk-network/web-wallet/issues/948
+[#953]: https://github.com/dusk-network/web-wallet/pull/953
 
 <!-- VERSIONS -->